Dedup session-bootstrap quiet-window const and bootstrap test
center and sidebar each defined an identical
const fullPaneCaptureQuietWindow = 2 * time.Second and an identical
TestCaptureExistingSessionBootstrap_RechecksExclusivityBeforeResize that
exercises common.CaptureExistingSessionBootstrap through their per-package
forwarders. Define the quiet window once as
common.FullPaneCaptureQuietWindow (both forwarders now reference it) and
move the bootstrap test into common's session_bootstrap_test.go, driving
common.CaptureExistingSessionBootstrap directly with an instrumented
SessionBootstrapFns; delete both per-package copies.

The per-package sessionBootstrapFns()/forwarders stay — they bind
package-private test-seam vars and are genuinely per-package.
