+++ internal/ui/sidebar/terminal_pty_attach.go
@@ -35,8 +35,6 @@ var (
 	verifyTerminalSessionTagsFn = verifyTerminalSessionTags
 )
 
-const fullPaneCaptureQuietWindow = 2 * time.Second
-
 type sessionBootstrapCapture = common.SessionBootstrapCapture
 
 func sessionBootstrapFns() common.SessionBootstrapFns {
@@ -55,7 +53,7 @@ func sessionBootstrapFns() common.SessionBootstrapFns {
 }
 
 func captureExistingSessionBootstrap(sessionName string, cols, rows int, opts tmux.Options) sessionBootstrapCapture {
-	return common.CaptureExistingSessionBootstrap(sessionName, cols, rows, fullPaneCaptureQuietWindow, opts, sessionBootstrapFns())
+	return common.CaptureExistingSessionBootstrap(sessionName, cols, rows, common.FullPaneCaptureQuietWindow, opts, sessionBootstrapFns())
 }
 
 func bootstrapSnapshotStillMatchesSession(sessionName string, bootstrap sessionBootstrapCapture, opts tmux.Options) bool {
